requirements.txt
streamlit==1.22.0
openai>=1.84.0
python-dotenv==1.0.0
opencv-python==4.7.0.72
pytesseract==0.3.10
Pillow==9.5.0
reportlab==3.6.12
boto3==1.28.38
numpy==1.24.3